Refactor BoatnetTrips component: move columns and data to props
